Fixing assests/images issues in deployed app + moved around catalouge files + moved favorites to the very right of navbar
